\name{rollshap-package}
\alias{rollshap-package}
\alias{rollshap}
\docType{package}
\title{
  \packageTitle{rollshap}
}
\description{
  \packageDescription{rollshap}
}
\details{
  \code{rollshap} is a package that provides analytical computation of rolling and expanding Shapley values for time-series data.
}
\author{
  \packageAuthor{rollshap}
}
\references{
  Shapley, L.S. (1953). "A Value for n-Person Games." In \emph{Contributions to the Theory of Games, Volume II}, edited by H.W. Kuhn and A.W. Tucker, 307-317. Princeton University Press. \doi{10.1515/9781400881970-018}
}
