\name{rollshap-package}
\alias{rollshap-package}
\alias{rollshap}
\docType{package}
\title{
  \packageTitle{rollshap}
}
\description{
  \packageDescription{rollshap}
}
\details{
  \code{rollshap} is a package that provides analytical computation of rolling Shapley values for time-series data.
}
\author{
  \packageAuthor{rollshap}
}